refactor: Product 관련 코드들을 entity로 이동
