+++ src/basic/entities/product/libs/index.ts
