+++ src/basic/entities/product/ui/ProductCard.tsx
