--- src/basic/features/product/types.ts
@@ -1,12 +1,0 @@
-export interface Product {
-  id: string;
-  name: string;
-  price: number;
-  stock: number;
-  discounts: Discount[];
-}
-
-export interface Discount {
-  quantity: number;
-  rate: number;
-}
